requirements.txt
Flask
azure-kusto-data
opencensus-ext-azure
